+++ themis/evaluation/pipelines/__init__.py
@@ -2,6 +2,7 @@
 
 from themis.evaluation.pipelines.composable_pipeline import (
     ComposableEvaluationPipeline,
+    ComposableEvaluationReportPipeline,
     EvaluationResult,
     EvaluationStep,
 )
@@ -10,6 +11,7 @@ from themis.evaluation.pipelines.standard_pipeline import EvaluationPipeline
 __all__ = [
     "EvaluationPipeline",
     "ComposableEvaluationPipeline",
+    "ComposableEvaluationReportPipeline",
     "EvaluationStep",
     "EvaluationResult",
 ]